Move property object distributon to `property.rs`
