+++ tests/graph/test-data/rust/src/seeding/distributions/mod.rs
@@ -1,7 +1,6 @@
 use error_stack::IntoReport;
 
 pub mod adaptors;
-pub mod entity;
 pub mod ontology;
 pub mod property;
 pub mod value;
